+++ lib/utils.dart
@@ -1,16 +1,180 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
+import 'package:medmap/const.dart';
+import 'package:medmap/models/r_profile.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:http/http.dart' as http;
 import 'dart:io';
 import 'package:path_provider/path_provider.dart';
 import 'package:path/path.dart';
-
+import 'dart:convert';
 import 'package:flutter/services.dart';
 import 'views/pdf_screen.dart';
 import 'package:share_plus/share_plus.dart';
 
 class Utils {
+  // start SP utils
+  static Future<void> setProfile(rProfile profile) async {
+    final prefs = await SharedPreferences.getInstance();
+    final jsonString = jsonEncode(profile.toJson());
+    await prefs.setString(Const.OBJ_PROFILE, jsonString);
+    // print('cekProfile stored: $jsonString');
+  }
+
+  static Future<rProfile?> getProfile() async {
+    final prefs = await SharedPreferences.getInstance();
+    final jsonString = prefs.getString(Const.OBJ_PROFILE);
+    if (jsonString != null) {
+      final userMap = jsonDecode(jsonString);
+      return rProfile.fromJson(userMap);
+    }
+    return null;
+  }
+
+  // Save a boolean value with the given key to SharedPreferences.
+  static Future<bool> setSpBool(String key, bool value) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      return prefs.setBool(key, value);
+    } catch (e) {
+      debugPrint("Error saving boolean: $e");
+      return false;
+    }
+  }
+
+  // Save an integer value with the given key to SharedPreferences.
+  static Future<bool> setSpInt(String key, int value) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      return prefs.setInt(key, value);
+    } catch (e) {
+      debugPrint("Error saving integer: $e");
+      return false;
+    }
+  }
+
+  // Save a double value with the given key to SharedPreferences.
+  static Future<bool> setSpDouble(String key, double value) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      return prefs.setDouble(key, value);
+    } catch (e) {
+      debugPrint("Error saving double: $e");
+      return false;
+    }
+  }
+
+  // Save a string value with the given key to SharedPreferences.
+  static Future<bool> setSpString(String key, String value) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      return prefs.setString(key, value);
+    } catch (e) {
+      debugPrint("Error saving string: $e");
+      return false;
+    }
+  }
+
+  // Save a list of strings with the given key to SharedPreferences.
+  static Future<bool> setSpStringList(String key, List<String> value) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      return prefs.setStringList(key, value);
+    } catch (e) {
+      debugPrint("Error saving string list: $e");
+      return false;
+    }
+  }
+
+  // Retrieve a boolean value with the given key from SharedPreferences.
+  static Future<bool?> getSpBool(String key) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      return prefs.getBool(key);
+    } catch (e) {
+      debugPrint("Error retrieving boolean: $e");
+      return null;
+    }
+  }
+
+  // Retrieve an integer value with the given key from SharedPreferences.
+  static Future<int?> getSpInt(String key) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      return prefs.getInt(key);
+    } catch (e) {
+      debugPrint("Error retrieving integer: $e");
+      return null;
+    }
+  }
+
+  // Retrieve a double value with the given key from SharedPreferences.
+  static Future<double?> getSpDouble(String key) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      return prefs.getDouble(key);
+    } catch (e) {
+      debugPrint("Error retrieving double: $e");
+      return null;
+    }
+  }
+
+  // Retrieve a string value with the given key from SharedPreferences.
+  static Future<String?> getSpString(String key) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      return prefs.getString(key);
+    } catch (e) {
+      debugPrint("Error retrieving string: $e");
+      return null;
+    }
+  }
+
+  // Retrieve a list of strings with the given key from SharedPreferences.
+  static Future<List<String>?> getSpStringList(String key) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      return prefs.getStringList(key);
+    } catch (e) {
+      debugPrint("Error retrieving string list: $e");
+      return null;
+    }
+  }
+
+  // get Loged in
+  static Future<bool> isUserLoggedIn() async {
+    final prefs = await getPreferencesInstance();
+    bool isLoggedIn = prefs.getBool(Const.IS_LOGED_IN) ?? false;
+    return isLoggedIn;
+  }
+
+  // Remove data associated with the given key from SharedPreferences.
+  static Future<void> removeSp(String key) async {
+    try {
+      final prefs = await getPreferencesInstance();
+      await prefs.remove(key);
+    } catch (e) {
+      debugPrint("Error removing data: $e");
+    }
+  }
+
+  // Clear all data from SharedPreferences.
+  static Future<void> clearSp() async {
+    try {
+      final prefs = await getPreferencesInstance();
+      await prefs.clear();
+    } catch (e) {
+      debugPrint("Error clearing data: $e");
+    }
+  }
+
+  // Get an instance of SharedPreferences.
+  static Future<SharedPreferences> getPreferencesInstance() async {
+    return await SharedPreferences.getInstance();
+  }
+  // end SP Utils
+
   // static void myLog(String message) {
   //   final pattern = RegExp('.{1,800}');
   //   pattern.allMatches(message).forEach((match) => print(match.group(0)));
@@ -188,5 +352,37 @@ class Utils {
     } else {
       return 'Not Available';
     }
+  }
+}
+
+abstract class StringValidator {
+  ValidationResult validate(String value);
+}
+
+RegExp _emailRegExp = RegExp(
+  r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+",
+);
+
+class ValidationResult {
+  ValidationResult({this.isValid = false, this.text});
+  final bool isValid;
+  final String? text;
+}
+
+class EmailValidator implements StringValidator {
+  EmailValidator(this.label);
+  final String label;
+
+  @override
+  ValidationResult validate(String value, {List<String>? targets}) {
+    if (value.isEmpty) {
+      return ValidationResult(text: 'Please enter $label');
+    }
+
+    if (!_emailRegExp.hasMatch(value)) {
+      return ValidationResult(text: 'Please enter the correct $label');
+    }
+
+    return ValidationResult(isValid: true);
   }
 }