+++ lib/views/submenu.dart
@@ -1,9 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:medmap/const.dart';
+import 'package:medmap/models/r_profile.dart';
+import 'package:medmap/utils.dart';
+import 'package:medmap/views/auth/login.dart';
+import 'package:medmap/views/dashboard.dart';
 import '../../app_localzations.dart';
 
 class Submenu extends StatefulWidget {
-  static const String route = '/products';
+  static const String route = '/submenu';
   Submenu({
     Key? key,
   }) : super(key: key);
@@ -13,11 +17,33 @@ class Submenu extends StatefulWidget {
 
 class _SubmenuState extends State<Submenu> {
   late String report;
+  rProfile? mProfile;
+  late String username = 'Username not set';
+
+  Future<String> getUsername() async {
+    var fetchedUsername = await Utils.getSpString(Const.USERNAME);
+    return fetchedUsername ?? 'Username not set';
+  }
+
   @override
   void initState() {
     super.initState();
-    Future.delayed(Duration.zero, () {
+    loadProfileData();
+  }
 
+Future<void> loadProfileData() async {
+    var profile = await Utils.getProfile();
+    // print('cek profile: $profile');
+    if (profile != null) {
+      setState(() {
+        mProfile = profile;
+      });
+    } else {
+      debugPrint('cek mProfile is null');
+    }
+    var fetchedUsername = await getUsername();
+    setState(() {
+      username = fetchedUsername;
     });
   }
 
@@ -41,32 +67,72 @@ class _SubmenuState extends State<Submenu> {
                     crossAxisAlignment: CrossAxisAlignment.start,
                     children: [
                       // SizedBox(height: 10),
-                      Text('Profile Name',
+                      Text(mProfile?.user?.username ?? username,
                           style: TextStyle(fontSize: 16)), // User's name
-                      Text('Company Name',
+                      Text(mProfile?.name ?? 'Company not set',
                           style: TextStyle(
                               fontSize: 14,
                               color: Colors.grey[600])), // User's role
                     ],
                   ),
-                  SizedBox(width: 10), // Space between the avatar and the text
+                  SizedBox(width: 10),
                   CircleAvatar(
-                    backgroundImage: NetworkImage(
-                        'https://example.com/profile_picture.jpg'), // Replace with your profile picture URL
+                    backgroundImage: NetworkImage(mProfile?.logo?.url ?? ''),
                     radius: 30, // Adjust the size of the circle avatar
                   ),
                 ],
               ),
             ),
+            IconButton(
+              icon: Icon(Icons.exit_to_app),
+              onPressed: () {
+                showDialog(
+                  context: context,
+                  builder: (BuildContext context) {
+                    return AlertDialog(
+                      title: Text('Exit'),
+                      content: Text('Are you sure you want to exit?'),
+                      actions: <Widget>[
+                        TextButton(
+                          child: Text('Cancel'),
+                          onPressed: () {
+                            Navigator.of(context).pop();
+                          },
+                        ),
+                        TextButton(
+                          child: Text('Exit'),
+                          onPressed: () async {
+                            await Utils.clearSp();
+                            await Navigator.pushAndRemoveUntil(
+                              context,
+                              MaterialPageRoute(
+                                  builder: (context) => Dashboard()),
+                              (Route<dynamic> route) => false,
+                            );
+                            // await Navigator.pushAndRemoveUntil(
+                            //   context,
+                            //   MaterialPageRoute(builder: (context) => Login()),
+                            // );
+                          },
+                        ),
+                      ],
+                    );
+                  },
+                );
+              },
+            ),
           ],
         ),
         backgroundColor: Colors.white, // Set the body background color to white
         body: CardsLayout(
-          report: AppLocalizations.of(context)!.translate('market_study_report'),
-          event: AppLocalizations.of(context)!.translate('product_launch_event'),
-          design: AppLocalizations.of(context)!.translate('product_brochure_design'),
+          report:
+              AppLocalizations.of(context)!.translate('market_study_report'),
+          event:
+              AppLocalizations.of(context)!.translate('product_launch_event'),
+          design: AppLocalizations.of(context)!
+              .translate('product_brochure_design'),
           policy: AppLocalizations.of(context)!.translate('privacy_policy'),
-          ),
+        ),
       ),
     );
   }
@@ -75,7 +141,11 @@ class _SubmenuState extends State<Submenu> {
 class CardsLayout extends StatelessWidget {
   final String report, event, design, policy;
 
-  CardsLayout({required this.report, required this.event, required this.design, required this.policy});
+  CardsLayout(
+      {required this.report,
+      required this.event,
+      required this.design,
+      required this.policy});
 
   @override
   Widget build(BuildContext context) {
@@ -177,33 +247,38 @@ class CardsLayout extends StatelessWidget {
               ),
             ),
           ),
-          Card(
-            shape: OutlineInputBorder(
-              borderRadius: BorderRadius.circular(15),
-              borderSide: BorderSide(color: Colors.black12, width: 1),
-            ),
-            color: Colors.white,
-            elevation: 0,
-            margin: EdgeInsets.symmetric(vertical: 20, horizontal: 20),
-            child: Center(
-              // Wrap the Column with Center to center its contents
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: <Widget>[
-                  Image.asset(
-                    Const.submenu_privacy,
-                    fit: BoxFit.contain,
-                    height: 30,
-                  ),
-                  SizedBox(height: 10),
-                  Padding(
-                    padding: EdgeInsets.symmetric(
-                        horizontal: 5.0), // Add padding to the left and right
-                    child: Text(policy,
-                        textAlign: TextAlign.center,
-                        style: TextStyle(fontSize: 10)),
-                  ),
-                ],
+          InkWell(
+            onTap: () {
+              Utils.launchURL(context, Const.URL_PRIVACY);
+            },
+            child: Card(
+              shape: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(15),
+                borderSide: BorderSide(color: Colors.black12, width: 1),
+              ),
+              color: Colors.white,
+              elevation: 0,
+              margin: EdgeInsets.symmetric(vertical: 20, horizontal: 20),
+              child: Center(
+                // Wrap the Column with Center to center its contents
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: <Widget>[
+                    Image.asset(
+                      Const.submenu_privacy,
+                      fit: BoxFit.contain,
+                      height: 30,
+                    ),
+                    SizedBox(height: 10),
+                    Padding(
+                      padding: EdgeInsets.symmetric(
+                          horizontal: 5.0), // Add padding to the left and right
+                      child: Text(policy,
+                          textAlign: TextAlign.center,
+                          style: TextStyle(fontSize: 10)),
+                    ),
+                  ],
+                ),
               ),
             ),
           ),